feature: ProfileCard 컴포넌트가 인자를 받도록 수정, Test 수정
