+++ src/component/page/BlogPage.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+import { useNavigate } from "react-router-dom";
 
 import ProfileCard from "../sidebar/ProfileCard";
 import CategoryList from "../sidebar/CategoryList"
@@ -11,16 +12,41 @@ const Sidebar = styled.div`
     align-items: center;
     width: 370px;
     min-height: 100%;
-    box-sizing: border-box;
     padding: 150px 0px;
     background-color: #064635;
 `;
 
+const buttonList = [
+    {
+        name: "Home",
+        url: "/blog"
+    },
+    {
+        name: "Info",
+        url: "#"
+    },
+    {
+        name: "Logout",
+        url: "#"
+    },
+    {
+        name: "write new",
+        url: "/blog/post"
+    },
+];
+
 function BlogPage(props) {
+    const navigate = useNavigate();
+
     return (
         <>
             <Sidebar>
-                <ProfileCard />
+                <ProfileCard
+                    nickname={`USERID`}
+                    imageSrc={`/image/defaultProfile.png`}
+                    imageClick={() => navigate("/profile")}
+                    buttonList={buttonList}
+                />
                 <CategoryList
                     list={categoryData}
                 />
