+++ src/test/Test.jsx
@@ -1,19 +1,14 @@
 import React from "react";
 import { BrowserRouter } from "react-router-dom";
-import styled from "styled-components";
 
 import ButtonTest from "../test/component/ui/ButtonTest"
 import ImageTest from "../test/component/ui/ImageTest";
-import ProfileCard from "../component/sidebar/ProfileCard";
+import ProfileCardTest from "./component/sidebar/ProfileCardTest";
 import CategoryTest from "./component/sidebar/CategoryTest";
-
-const Wrapper = styled.div`
-    width: 400px;
-    padding: 10px;
-    background-color:#064635;
-`
+import BlogPage from "../component/page/BlogPage";
 
 function Test(props) {
+
     return (
         <BrowserRouter>
             <ButtonTest />
@@ -21,12 +16,9 @@ function Test(props) {
             <ImageTest />
             <hr/>
             <h3>Sidebar TEST</h3>
-            <p>ProfileCard</p>
-            <ProfileCard />
-            <br/>
-            <Wrapper>
-                <CategoryTest/>
-            </Wrapper>
+            <ProfileCardTest />
+            <CategoryTest/>
+            {/* <BlogPage /> */}
         </BrowserRouter>
     );
 }
